requirements.txt
Django
djangorestframework
psycopg2-binary
opensearch-py
python-dotenv
gunicorn
django-prometheus
prometheus-client
debugpy